Make the allowed paths less strict
The current regex is too strict with regards to username and repository name. For example, the selfie button does not show up for `https://github.com/github/scientist.net/issues/new` because of the `.` in `scientist.net`. Also, the regex would try to match sites such as "foogithub.com". Since GitHub requires https and the browser redirects URLs with backslashes to those with forward slashes, it's safe to simplify the regex as I've done here and include the `https://` portion in the regex.